ANDROID: cpufreq: Remove seq_printf from critical path
Calls to uid_time_in_state, uid_concurrent_*_time are
on the critical path when the screen changes. seq_printf
is taking a significant portion of the time. Replace
seq_printf with the faster seq_write function.

Since the files will not be human readable, move them
to a seperate directory. Leave a human readable version
of uid_time_in_state where it is for debugging and
testing purposes.

b/71718257

Change-Id: Ie947dd0caa26682a655a164b195ad72fd6693c1b
Signed-off-by: Marissa Wall <marissaw@google.com>
Signed-off-by: utsavbalar1231 <utsavbalar1231@gmail.com>