msm: kgsl: Remove POPP
POPP constantly attempts to lower the GPU's frequency behind the governor's
back in order to save power; however, the GPU governor in use
(msm-adreno-tz) is very good at determining the GPU's load and selecting
an appropriate frequency to run the GPU at.

POPP was created long ago (perhaps when msm-adreno-tz didn't exist or
didn't work so well), so it is clearly deprecated. Remove it.

Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>
Signed-off-by: Raphiel Rollerscaperers <raphielscape@outlook.com>
[krasCGQ: Remove default_pwrlevel from adreno.c to fix compilation]
Signed-off-by: Albert I <krascgq@outlook.co.id>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>
Signed-off-by: utsavbalar1231 <utsavbalar1231@gmail.com>